type-double commit

+++ src/main/java/tp10/metier/ActionCompliquee.java
@@ -38,16 +38,16 @@ public class ActionCompliquee extends Action {
     
     @Override
     public float valeur(Jour j) {
-        float total;
+        double total;
         
         total = 0;
         
          // parcours des clefs
         for(ActionSimple act : this.tblComposition.keySet()) {
             total = total + 
-                    (act.valeur(j) * this.tblComposition.get(act).getPourcentage());
+                    ((double)act.valeur(j) * (double)this.tblComposition.get(act).getPourcentage());
         }
-        return total;
+        return (float)total;
     }
     
     public void enrgProportion(ActionSimple as, float pourcentage) {
